fixed light sensor return value

+++ include/LightSensor.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void setupLightSensor(uint8_t cEN, uint8_t scl, uint8_t sda);
-unsigned int getLightValue();
+float getLightValue();
