+++ src/LightSensor.cpp
@@ -7,6 +7,6 @@ void setupLightSensor(uint8_t cEN, uint8_t scl, uint8_t sda) {
     myLux.begin();
 }
 
-unsigned int getLightValue() {
+float getLightValue() {
     return myLux.lightStrengthLux();
 }
